agent_company/requirements.txt
openai>=1.14.0
langchain>=0.2.0
faiss-cpu
duckduckgo-search
pytest
python-dotenv 
# --- connectomics tracing agent dependencies ---
cloud-volume>=9.10.0           # streaming EM data
tensorstore>=0.1.45            # high‑throughput I/O (optional fallback)
numpy>=1.21.0                  # core numeric ops
dask[array]>=2024.2.1          # distributed metric computation
prometheus-client>=0.18.0      # telemetry metrics endpoint
torch>=1.12.0                   # backbone for FFN‑v2 & LoRA adapters
torchvision>=0.13.0
peft>=0.10.0                   # parameter‑efficient fine‑tuning (LoRA)
accelerate>=0.27.0             # lightweight training loop helper
# --- additional tracing agent dependencies ---
scipy>=1.7.0                  # scientific computing (proofreading)
psutil>=5.9.0                  # system monitoring
transformers>=4.37.0           # Hugging Face transformers
google-cloud-firestore>=2.8.0 # Firestore integration
h5py>=3.7.0                   # HDF5 file support
tensorflow>=2.8.0             # TensorFlow for FFN-v2 (optional)

# Core dependencies
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Deep Learning
torch>=1.12.0
torchvision>=0.13.0
tensorflow>=2.8.0
tensorflow-gpu>=2.8.0

# Distributed Computing
dask>=2022.1.0
dask[distributed]>=2022.1.0
distributed>=2022.1.0

# Large-scale Data Storage
zarr>=2.12.0
numcodecs>=0.10.0
h5py>=3.7.0
cloudvolume>=3.0.0

# Cloud Storage
google-cloud-firestore>=2.8.0
google-cloud-storage>=2.5.0
boto3>=1.24.0

# Monitoring and Telemetry
prometheus-client>=0.14.0
psutil>=5.9.0
GPUtil>=1.4.0

# Web Framework
fastapi>=0.78.0
uvicorn>=0.18.0
pydantic>=1.9.0

# Database
sqlalchemy>=1.4.0
alembic>=1.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=3.0.0

# Development
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Production
gunicorn>=20.1.0
supervisor>=4.2.0
redis>=4.3.0
celery>=5.2.0

# Image Processing
opencv-python>=4.6.0
Pillow>=9.0.0
imageio>=2.19.0

# Scientific Computing
numba>=0.56.0
numba[llvmlite]>=0.56.0
cupy-cuda11x>=10.0.0

# Memory Management
memory-profiler>=0.60.0
pympler>=0.9.0

# Logging and Monitoring
structlog>=21.5.0
sentry-sdk>=1.9.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Networking
aiohttp>=3.8.0
requests>=2.28.0

# Utilities
tqdm>=4.64.0
click>=8.1.0
rich>=12.0.0